feat(dataflow): bundle TypeScript client + agent recipe in the wheel
Ship ``marimo/_dataflow/clients/typescript/dataflow.tsx`` and
``marimo/_dataflow/clients/AGENT.md`` inside the wheel and surface them
through a new ``marimo dataflow`` CLI:

    marimo dataflow client > src/dataflow.tsx     # vendor the TS lib
    marimo dataflow agent                          # print agent recipe
    marimo dataflow {client,agent} --path          # filesystem path

The ``--path`` flag lets agent skill loaders register the bundled file
directly without copying contents around. Anything with marimo installed
gets the dataflow client and the agent recipe for free.

The demo's ``frontend/src/dataflow.tsx`` becomes a relative symlink into
the canonical source-of-truth path, eliminating duplication.

Adds ``docs/guides/dataflow_api.md`` (wired into the mkdocs nav) as the
public reference. Trims the demo README to point at the new CLI and
flags the symlink + Windows caveat.

Also sweeps lint + mypy across this branch's earlier commits:

- TC002/TC003 imports → ``TYPE_CHECKING`` blocks
- ``_SchemaListener`` properly subclasses ``SessionConsumer``
- ``_broadcast_values_for_run`` accepts ``AbstractSet[str] | None``
- Annotate ``event_stream`` and ``_to_dict`` return values
- Replace ``session.room.consumers`` with the public ``session.consumers``

Made-with: Cursor
