+++ mkdocs.yml
@@ -134,6 +134,7 @@ nav:
       - Using your own editor: guides/editor_features/watching.md
       - Run notebooks as apps: guides/apps.md
       - Run notebooks as scripts: guides/scripts.md
+      - Run as a dataflow API (experimental): guides/dataflow_api.md
       - Reuse functions and classes: guides/reusing_functions.md
       - Run in the cloud with molab: guides/molab.md
       - Export to other formats:
