+++ tests/_cli/test_cli_dataflow.py
@@ -1,0 +1,69 @@
+# Copyright 2026 Marimo. All rights reserved.
+"""Smoke tests for the ``marimo dataflow`` CLI subcommands.
+
+These verify both that the bundled assets are reachable through the
+installed package's ``importlib.resources`` view *and* that the click
+plumbing is wired up. They intentionally don't pin the asset contents —
+those are exercised by the hashing test below, which catches accidental
+import drift between the wheeled file and the source-of-truth path.
+"""
+
+from __future__ import annotations
+
+import hashlib
+import subprocess
+from pathlib import Path
+
+import pytest
+
+from marimo._utils.paths import marimo_package_path
+from marimo._utils.platform import is_windows
+
+
+def _bundled(*parts: str) -> Path:
+    p = marimo_package_path()
+    for part in parts:
+        p = p / part
+    return Path(str(p))
+
+
+@pytest.mark.xfail(condition=is_windows(), reason="flaky on Windows")
+@pytest.mark.parametrize(
+    ("subcommand", "needle"),
+    [
+        ("client", "DataflowProvider"),
+        ("agent", "Dataflow API — Agent Recipe"),
+    ],
+)
+def test_dataflow_dump_to_stdout(subcommand: str, needle: str) -> None:
+    p = subprocess.run(
+        ["marimo", "dataflow", subcommand],
+        capture_output=True,
+    )
+    assert p.returncode == 0, p.stderr.decode()
+    assert needle in p.stdout.decode()
+
+
+@pytest.mark.xfail(condition=is_windows(), reason="flaky on Windows")
+@pytest.mark.parametrize(
+    ("subcommand", "asset_parts"),
+    [
+        ("client", ("_dataflow", "clients", "typescript", "dataflow.tsx")),
+        ("agent", ("_dataflow", "clients", "AGENT.md")),
+    ],
+)
+def test_dataflow_path_flag(
+    subcommand: str, asset_parts: tuple[str, ...]
+) -> None:
+    """``--path`` resolves to the bundled file and the bytes match."""
+    p = subprocess.run(
+        ["marimo", "dataflow", subcommand, "--path"],
+        capture_output=True,
+    )
+    assert p.returncode == 0, p.stderr.decode()
+    reported = Path(p.stdout.decode().strip())
+    assert reported.exists(), reported
+    expected_bytes = _bundled(*asset_parts).read_bytes()
+    assert reported.read_bytes() == expected_bytes
+    # Sanity check: the contents are stable enough for downstream hashing.
+    assert hashlib.sha256(expected_bytes).hexdigest()
